feat(harmonyos-sans-sc-vf): add 122 new woff2 font files
This commit introduces a total of 122 new woff2 font files for the HarmonyOS Sans SC VF
font family. The files are added to the `harmonyos-sans-sc-vf` directory and include a
range of weights and styles to support the extended character set and design variations.

Each file is uniquely identified by a combination of alphanumeric characters and has a
`.woff2` file extension. The addition of these files enhances the font coverage and
provides more options for typography within the HarmonyOS ecosystem.